add drink water logic and reset if new day add hive for local storage

+++ lib/feature/drink_water/presentation/views/widgets/percent_indicator_body.dart
@@ -11,8 +11,8 @@ class PercentIndicatorBody extends StatelessWidget {
     this.onTap,
   });
 
-  final double currentIntake;
-  final double dailyTarget;
+  final int currentIntake;
+  final int dailyTarget;
   final Function()? onTap;
 
   @override
