+++ lib/feature/drink_water/presentation/views/widgets/percent_indicator_widget.dart
@@ -14,8 +14,8 @@ class PercentIndicatorWidget extends StatelessWidget {
   });
 
   final double percent;
-  final double currentIntake;
-  final double dailyTarget;
+  final int currentIntake;
+  final int dailyTarget;
   final Function()? onTap;
 
   @override
